Alternate main window with a couple of fixes to the one in the dev branch but has scrolling problem
Has focused text on text fields now. Opens a little window when setup button is pressed. The only problem is that the scrolling has become limited so you can't see all the row entries you add. I think the reason is due to the sizing of the window and the panels.